Replacing `Instagram` with `InstagramGraph` and fixed CS errors
