+++ src/Provider/Instagram.php
@@ -2,30 +2,30 @@
 /*!
 * Hybridauth
 * https://hybridauth.github.io | https://github.com/hybridauth/hybridauth
-*  (c) 2017 Hybridauth authors | https://hybridauth.github.io/license.html
+*  (c) 2020 Hybridauth authors | https://hybridauth.github.io/license.html
 */
 
 namespace Hybridauth\Provider;
 
 use Hybridauth\Adapter\OAuth2;
+use Hybridauth\Data\Collection;
 use Hybridauth\Exception\UnexpectedApiResponseException;
-use Hybridauth\Data;
 use Hybridauth\User;
 
 /**
- * Instagram OAuth2 provider adapter.
+ * Instagram OAuth2 provider adapter via Facebook Graph API.
  */
 class Instagram extends OAuth2
 {
     /**
      * {@inheritdoc}
      */
-    protected $scope = 'basic';
+    protected $scope = 'user_profile,user_media';
 
     /**
      * {@inheritdoc}
      */
-    protected $apiBaseUrl = 'https://api.instagram.com/v1/';
+    protected $apiBaseUrl = 'https://graph.instagram.com/';
 
     /**
      * {@inheritdoc}
@@ -35,7 +35,7 @@ class Instagram extends OAuth2
     /**
      * {@inheritdoc}
      */
-    protected $accessTokenUrl = 'https://api.instagram.com/oauth/access_token';
+    protected $accessTokenUrl = 'https://api.instagram.com/oauth/access_token/';
 
     /**
      * {@inheritdoc}
@@ -45,43 +45,88 @@ class Instagram extends OAuth2
     /**
      * {@inheritdoc}
      */
-    protected function initialize()
-    {
-        parent::initialize();
-
-        // The Instagram API requires an access_token from authenticated users
-        // for each endpoint, see https://www.instagram.com/developer/endpoints.
-        $accessToken = $this->getStoredData($this->accessTokenName);
-        $this->apiRequestParameters[$this->accessTokenName] = $accessToken;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function getUserProfile()
     {
-        $response = $this->apiRequest('users/self/');
+        $parameters = [
+            'fields' => 'id,username,account_type,media_count',
+            $this->accessTokenName => $this->getStoredData($this->accessTokenName),
+        ];
 
-        $data = new Data\Collection($response);
+        $response = $this->apiRequest('me', 'GET', $parameters);
 
-        if (! $data->exists('data')) {
+        $data = new Collection($response);
+
+        if (!$data->exists('id')) {
             throw new UnexpectedApiResponseException('Provider API returned an unexpected response.');
         }
 
         $userProfile = new User\Profile();
 
-        $data = $data->filter('data');
-
-        $userProfile->identifier  = $data->get('id');
-        $userProfile->description = $data->get('bio');
-        $userProfile->photoURL    = $data->get('profile_picture');
-        $userProfile->webSiteURL  = $data->get('website');
-        $userProfile->displayName = $data->get('full_name');
-        $userProfile->displayName = $userProfile->displayName ?: $data->get('username');
+        $userProfile->identifier = $data->get('id');
+        $userProfile->displayName = $data->get('username');
         $userProfile->profileURL = "https://instagram.com/{$data->get('username')}";
-        
-        $userProfile->data = (array) $data->get('counts');
+        $userProfile->data = [
+            'account_type' => $data->get('account_type'),
+            'media_count' => $data->get('media_count'),
+        ];
 
         return $userProfile;
+    }
+
+    /**
+     * Fetch user medias.
+     *
+     * @param int $limit Number of elements per page.
+     * @param string $pageId Current pager ID.
+     * @param array|null $fields Fields to fetch per media.
+     *
+     * @throws \Hybridauth\Exception\UnexpectedApiResponseException If API's response is not valid.
+     *
+     * @return \Hybridauth\Data\Collection Response encapsulated in a `Collection`.
+     */
+    public function getUserMedia(int $limit = 12, ?string $pageId = null, array $fields = null)
+    {
+        if ($fields === null || count($fields) == 0) {
+            $fields = [
+                'id', 'caption', 'media_type', 'media_url',
+                'thumbnail_url', 'permalink', 'timestamp', 'username'
+            ];
+        }
+
+        $fields = implode(',', $fields);
+
+        $pager = $pageId !== null ? '&after=' . $pageId : '';
+        $response = $this->apiRequest('me/media?fields=' . $fields . '&limit=' . $limit . $pager);
+
+        $data = new Collection($response);
+        if (!$data->exists('data')) {
+            throw new UnexpectedApiResponseException('Provider API returned an unexpected response.');
+        }
+
+        return $data;
+    }
+
+    /**
+     * Fetches a single user's media.
+     *
+     * @param string $mediaId Media ID.
+     * @param array|null $fields Fields to fetch per media.
+     *
+     * @return object Raw response.
+     */
+    public function getMedia($mediaId, array $fields = null)
+    {
+        if ($fields === null || count($fields) == 0) {
+            $fields = [
+                'id', 'caption', 'media_type', 'media_url',
+                'thumbnail_url', 'permalink', 'timestamp', 'username'
+            ];
+        }
+
+        $fields = implode(',', $fields);
+
+        $response = $this->apiRequest($mediaId . '?fields=' . $fields);
+
+        return $response;
     }
 }
